Splitting datasets, descriptions, and tests up by modality. Note: tests and descriptions are broken, and local data files are missing.
