+++ nidata/_external/nibabel102/doc/source/api.rst
@@ -59,4 +59,4 @@ Miscellaneous Helpers
    fileholders
    filename_parser
    onetime
-   tmpdirs
+   get_tmpdirs
