+++ nidata/atlas/description/craddock_2012.rst
