+++ nidata/localizer/description/brainomics_localizer.rst
