+++ nidata/resting_state/description/adhd.rst
