+++ nidata/resting_state/description/nyu.rst
