+++ nidata/resting_state/description/smith_2009.rst
